+++ cli/cli.hs
@@ -102,6 +102,7 @@ data CommonOptions = CommonOptions
   , cacheDir      ::Maybe String
   , earlyAbort    ::Bool
   , mergeMaxBudget :: Int
+  , maxDynSize    ::Int
   }
 
 commonOptions :: Parser CommonOptions
@@ -133,6 +134,7 @@ commonOptions = CommonOptions
   <*> (optional $ strOption $ long "cache-dir" <> help "Directory to save and load RPC cache")
   <*> (switch $  long "early-abort" <> help "Stop exploration immediately upon finding the first counterexample")
   <*> (option auto $ long "merge-max-budget" <> showDefault <> value 100 <> help "Max instructions for speculative merge exploration during path merging")
+  <*> (option auto $ long "max-dyn-size" <> showDefault <> value 64 <> help "Max byte length for concretized dynamic types (bytes, string) in symbolic arguments")
 
 data CommonExecOptions = CommonExecOptions
   { address       ::Maybe Addr
@@ -377,6 +379,7 @@ main = do
         , onlyDeployed = cOpts.onlyDeployed
         , earlyAbort = cOpts.earlyAbort
         , mergeMaxBudget = cOpts.mergeMaxBudget
+        , maxDynSize = cOpts.maxDynSize
         } }
 
 
@@ -416,7 +419,7 @@ equivalence eqOpts cOpts = do
                             , loopHeuristic = cOpts.loopDetectionHeuristic
                             }
                           }
-  calldata <- buildCalldata cOpts eqOpts.sig eqOpts.arg
+  (calldata, _caveats) <- buildCalldata cOpts eqOpts.sig eqOpts.arg
   solver <- liftIO $ getSolver cOpts.solver
   cores <- liftIO $ unsafeInto <$> getNumProcessors
   let solverCount = fromMaybe cores cOpts.numSolvers
@@ -477,8 +480,9 @@ getRoot cmd = maybe getCurrentDirectory makeAbsolute cmd.root
 parseMaxIters :: Integer -> Maybe Integer
 parseMaxIters num = if num < 0 then Nothing else Just num
 
--- | Builds a buffer representing calldata based on the given cli arguments
-buildCalldata :: App m => CommonOptions -> Maybe Text -> [String] -> m (Expr Buf, [Prop])
+-- | Builds a buffer representing calldata based on the given cli arguments,
+-- together with any program-wide soundness caveats incurred (see 'Caveat').
+buildCalldata :: App m => CommonOptions -> Maybe Text -> [String] -> m ((Expr Buf, [Prop]), [Caveat])
 buildCalldata cOpts sig arg = case (cOpts.calldata, sig) of
   -- fully abstract calldata
   (Nothing, Nothing) -> mkCalldata Nothing []
@@ -488,7 +492,7 @@ buildCalldata cOpts sig arg = case (cOpts.calldata, sig) of
     if (isNothing val) then liftIO $ do
       putStrLn $ "Error, invalid calldata: " <>  show c
       exitFailure
-    else pure (ConcreteBuf (fromJust val), [])
+    else pure ((ConcreteBuf (fromJust val), []), [])
   -- calldata according to given abi with possible specializations from the `arg` list
   (Nothing, Just sig') -> do
     method' <- liftIO $ functionAbi sig'
@@ -505,7 +509,7 @@ symbCheck cFileOpts sOpts cExecOpts cOpts = do
   let block' = maybe Fetch.Latest Fetch.BlockNumber cExecOpts.block
       blockUrlInfo = (,) block' <$> cExecOpts.rpc
   sess <- Fetch.mkSession cOpts.cacheDir cExecOpts.block
-  calldata <- buildCalldata cOpts sOpts.sig sOpts.arg
+  (calldata, caveats) <- buildCalldata cOpts sOpts.sig sOpts.arg
   preState <- symvmFromCommand cExecOpts sOpts cFileOpts sess calldata
   errCodes <- case sOpts.assertions of
     Nothing -> pure defaultPanicCodes
@@ -534,6 +538,7 @@ symbCheck cFileOpts sOpts cExecOpts cOpts = do
     case res of
       [] -> do
         liftIO $ putStrLn "\nQED: No reachable property violations discovered\n"
+        liftIO $ printCaveats caveats
         showExtras solvers sOpts calldata expr
       _ -> do
         let cexs = snd <$> mapMaybe getCex res
@@ -546,6 +551,7 @@ symbCheck cFileOpts sOpts cExecOpts cOpts = do
                  ] <> fmap (formatCex (fst calldata) Nothing) cexs
         liftIO $ T.putStrLn $ T.unlines counterexamples
         liftIO $ printWarnings Nothing mempty expr res "symbolically"
+        liftIO $ printCaveats caveats
         showExtras solvers sOpts calldata expr
         liftIO exitFailure